# api/openapi.yaml
openapi: 3.1.0
info:
  title: sendplane API
  version: 0.1.0
  summary: Mail sending engine REST API.
  description: |
    sendplane is a library-first mail sending engine (see docs/architecture.md).
    The host application owns accounts, RBAC and the unsubscribe destination;
    sendplane owns templates, campaigns, deliveries, bounces and tracking.

    ## Mounting

    `Sendplane.Handler()` serves everything below `/api/v1` plus the public
    tracking routes below `/t`. The host mounts that handler on its own mux, so
    the origin in `servers` is the host's.

    ## Tenancy

    A tenant is never a path parameter. The host's `TenantResolver` derives it
    from the authenticated principal or the request, and the store layer is
    bound to it (ADR-0001, ADR-0006). Every object in this API therefore belongs
    to the caller's tenant implicitly.

    ## Authentication

    Authentication is host-provided. The reference binary ships an API-key and a
    JWT `Authenticator`, so the spec declares a generic `bearerAuth` and
    `apiKeyAuth` scheme; a host that plugs in its own `Authenticator` may accept
    something else entirely. Both schemes are optional at the document level and
    the public tracking routes clear `security` explicitly.

    ## Authorization

    Every operation declares the closed-set `Action` it requires through the
    `x-sendplane-action` extension (architecture 3). The host's `Authorizer`
...
    `Authorize` because it is public.

...
    | --- | --- |
...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /probe/inbound/{provider}:
    post:
      operationId: probeInbound
      summary: Delivered probe mail from an inbound webhook provider
      description: |
        Completes a pending `ProbeRun` from a message that was forwarded here
        after delivery, for probe mailboxes of `kind: webhook` (ADR-0016).

        The route is **mounted only for the providers the deployment
        configured** (`probe.webhooks` in the host config), and `{provider}`
        is that provider's name. A deployment may move a provider to a path of
        its own, so treat this as the default rather than as a fixed URL. An
        unconfigured provider answers `404`.

        The format sendplane defines and registers by default is `sendplane`:
        body `SendplaneInboundMessage`, authenticated with
        `X-Sendplane-Signature`. Any other provider is that provider's own
        body and its own signature header — the schema below does not apply to
        it, and the payload is passed through verbatim.

        `X-Sendplane-Signature: t=<unix seconds>,v1=<hex>[,v1=<hex>...]`,
        where each `v1` is `HMAC-SHA256(secret, "<t>" + "." + <raw body>)`,
        hex. Any `v1` matching any configured secret is accepted, which is how
        a secret is rotated without a window where deliveries are refused, and
        a `t` more than the configured tolerance (default 300s) from the
        server's clock is refused as a replay. The signature covers the raw
        bytes, so it is checked before the body is parsed.

        Whatever forwards the mail **must preserve the
        `Authentication-Results`, `Received` and `X-Sendplane-Probe`
        headers**: the first two are the whole evidence a verdict is made of,
        and the third is what attributes the delivery to a tenant.

        Redelivery is safe. A run that is already finished is a no-op, and so
        is a byte-identical body that was already handled.
      tags: [Inbound (public)]
      security: []
      x-sendplane-action: none
      parameters:
        - name: provider
          in: path
          required: true
          description: Registered inbound provider name.
          schema: { type: string, example: sendplane }
      requestBody:
        required: true
        description: |
          For the `sendplane` provider, a `SendplaneInboundMessage`. For any
          other provider, that provider's own payload, verbatim. Either way
          the raw bytes are what the signature covers.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendplaneInboundMessage'
      responses:
        '200':
          description: |
            Recorded, already recorded, or not a sendplane probe. The body is a
            one-word outcome (`accepted`, `ignored`) for a human reading logs.
            "Not ours" is a `200` on purpose: a `4xx` would make a sender retry
            forever a message sendplane will never want.
          content:
            text/plain:
              schema: { type: string }
        '400':
          description: |
            The signature verified but the body is not the provider's format.
            Retrying the same bytes cannot help, so this is not a deferral.
        '401':
          description: |
            The signature is missing, malformed, made with an unknown secret,
            or carries a timestamp outside the tolerance.
        '404':
          description: No inbound webhook is configured for this provider.
        '413':
          description: The body is over the 1 MiB inbound cap.
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503':
          description: |
            Could not be recorded now; the sender should hold the mail and
            retry. `Retry-After` says when.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Credential the host's `Authenticator` accepts. The reference binary
        reads a JWT validated against a JWKS; another host may accept anything
        it likes.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Static API key. The reference binary maps a key to a principal and a
        tenant; the tenant is never sent by the client.
...
  responses:
...
    TooManyRequests:
      description: Rate limited (`rate_limited`).
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema: { type: integer, format: int32 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      description: The single error shape every non-2xx response uses.
      required: [code, message]
      properties:
        code:
          description: |
            Stable machine-readable code. Clients switch on this, never on the
            message.
          type: string
          enum:
            - invalid_request
            - invalid_cursor
            - unauthenticated
            - forbidden
            - not_found
            - duplicate
            - version_conflict
            - invalid_state
            - payload_too_large
            - validation_failed
            - limit_exceeded
            - missing_i18n_keys
            - template_not_published
            - render_failed
            - precondition_failed
            - rate_limited
            - platform_read_only
            - sender_use_denied
            - tenant_vars_missing
            - from_domain_not_owned
            - transport_not_assignable
            - internal
        message:
          type: string
          description: Human-readable English explanation; not for display to end users.
        details:
          type: array
          description: Per-field or per-item detail, when the failure has any.
          items: { $ref: '#/components/schemas/ErrorDetail' }
      examples:
        - code: version_conflict
          message: 'template 018f...: version 4 is stale, current is 5'
    ErrorDetail:
      type: object
      description: One field-level or item-level problem.
      required: [message]
      properties:
        field:
          type: string
          description: JSON pointer or dotted path into the request body.
        message: { type: string }
        value:
          description: The offending value, when it is safe to echo.
...
    SendplaneInboundMessage:
      type: object
      description: |
        The body of the `sendplane` inbound webhook format (ADR-0016) — the
        smallest payload that carries what a probe verdict reads. Anything that
        can POST JSON can produce it: a mail-receiving service's webhook, an
        email worker, a script on an MX.

        Only `from` and `headers` decide anything. Unknown fields are ignored,
        so the format can grow; a body over 1 MiB is refused.
      required: [from]
      properties:
        from:
          type: string
          description: Envelope sender. Required - a delivery without one is not a mail.
          example: news@example.com
        to:
          type: array
          description: Envelope recipients. Recorded, not read by the verdict.
          items: { type: string }
        headers:
          type: object
          description: |
            The message's header block, `{"Header-Name": ["value", ...]}`. Names
            are matched case-insensitively; the **order of the values within one
            name** is the contract, because the `Received` chain is read
            bottom-up and the *first* `Authentication-Results` is a trust
            decision.

            `Authentication-Results`, `Received` and `X-Sendplane-Probe` must
            all survive the forward. Without the first two there is no evidence
            to judge; without the third the delivery cannot be attributed to a
            tenant and is answered `200 ignored`.
          additionalProperties:
            type: array
            items: { type: string }
        text:
          type: string
          description: Plain-text body. Optional; the verdict is made of headers.
      example:
        from: news@example.com
        to: [probe@example.net]
        headers:
          Authentication-Results: ['mx.example.net; spf=pass; dkim=pass header.d=example.com; dmarc=pass']
          Received: ['from mx.example.net by inbox.example.net; Tue, 22 Sep 2026 12:00:30 +0000']
          X-Sendplane-Probe: ['acme/01JB.../9f2c...']
        text: "sendplane loopback health probe.\r\n"
...